Refactoring CreatePet
Refactoring du composant FormPet et du composant CreatePet car ils ne
respectaient pas la structure de fichiers mise en place
Le composant FormPet est donc supprimé car il ne sera utilisé qu'une
seule fois et le contenu de celui-ci est ajouté au composant CreatePet
