--- Web App Frontend/src/Components/FormPet/FormPet.test.js
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-import FormPet from './FormPet';
-
-describe('FormPet', () => {
-    it('Should render without crash', async () => {
-        render(
-            <FormPet/>
-        )
-    })
-})
